refactor: improve error message and fix some error
 - two packages with the same version but different channels are not
 - use type 'std::optional<std::string>' for subRef of UAB

+++ libs/linglong/src/linglong/builder/linglong_builder.cpp
@@ -785,10 +785,17 @@ set -e
                       2);
     qDebug() << "import binrary to layers";
     package::LayerDir binaryOutputLayerDir = binaryOutput.absoluteFilePath("..");
-    result = this->repo.remove(*ref);
-    if (!result) {
-        qWarning() << "remove" << ref->toString() << result.error().message();
+
+    // remove ref if exist
+    auto layerDir = this->repo.getLayerDir(*ref);
+    if (layerDir) {
+        auto result = this->repo.remove(*ref);
+        if (!result) {
+            return LINGLONG_ERR("failed to remove" + ref->toString() + ":"
+                                + result.error().message());
+        }
     }
+
     auto localLayer = this->repo.importLayerDir(binaryOutputLayerDir);
     if (!localLayer) {
         return LINGLONG_ERR(localLayer);
@@ -854,10 +861,8 @@ utils::error::Result<void> Builder::exportUAB(const QString &destination, const 
         packager.include(project.include.value());
     }
 
-    auto baseRef = pullDependency(QString::fromStdString(this->project.base),
-                                  this->repo,
-                                  "binary",
-                                  cfg.skipPullDepend.has_value() && *cfg.skipPullDepend);
+    auto baseRef =
+      pullDependency(QString::fromStdString(this->project.base), this->repo, "binary", false);
     if (!baseRef) {
         return LINGLONG_ERR(baseRef);
     }
@@ -871,7 +876,7 @@ utils::error::Result<void> Builder::exportUAB(const QString &destination, const 
         auto ref = pullDependency(QString::fromStdString(*this->project.runtime),
                                   this->repo,
                                   "binary",
-                                  cfg.skipPullDepend.has_value() && *cfg.skipPullDepend);
+                                  false);
         if (!ref) {
             return LINGLONG_ERR(ref);
         }
